// ui/src/lib/api-client/models/PublicIdentityProvider.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Proxy Smart
 * SMART on FHIR Proxy + Healthcare Administration API using Keycloak and Elysia
 *
 * The version of the OpenAPI document: 0.0.1-RELEASE.202510170109.139bb12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PublicIdentityProvider
 */
export interface PublicIdentityProvider {
    /**
     * Provider alias
     * @type {string}
     * @memberof PublicIdentityProvider
     */
    alias: string;
    /**
     * Provider type
     * @type {string}
     * @memberof PublicIdentityProvider
     */
    providerId: string;
    /**
     * Display name
     * @type {string}
     * @memberof PublicIdentityProvider
     */
    displayName: string;
    /**
     * Whether provider is enabled
     * @type {boolean}
     * @memberof PublicIdentityProvider
     */
    enabled: boolean;
}

/**
 * Check if a given object implements the PublicIdentityProvider interface.
 */
export function instanceOfPublicIdentityProvider(value: object): value is PublicIdentityProvider {
    if (!('alias' in value) || value['alias'] === undefined) return false;
    if (!('providerId' in value) || value['providerId'] === undefined) return false;
    if (!('displayName' in value) || value['displayName'] === undefined) return false;
    if (!('enabled' in value) || value['enabled'] === undefined) return false;
    return true;
}

export function PublicIdentityProviderFromJSON(json: any): PublicIdentityProvider {
    return PublicIdentityProviderFromJSONTyped(json, false);
}

export function PublicIdentityProviderFromJSONTyped(json: any, ignoreDiscriminator: boolean): PublicIdentityProvider {
    if (json == null) {
        return json;
    }
    return {
        
        'alias': json['alias'],
        'providerId': json['providerId'],
        'displayName': json['displayName'],
        'enabled': json['enabled'],
    };
}

export function PublicIdentityProviderToJSON(json: any): PublicIdentityProvider {
    return PublicIdentityProviderToJSONTyped(json, false);
}

export function PublicIdentityProviderToJSONTyped(value?: PublicIdentityProvider | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'alias': value['alias'],
        'providerId': value['providerId'],
        'displayName': value['displayName'],
        'enabled': value['enabled'],
    };
}